Add files via upload
This is a MATLAB script apropos complex baseband architecture. Intermediate frequency(IF) is generated by multiplication of received signal with complex phasor to get I and Q symbols. FFT is taken of the input array of the form I+jQ followed by modulus to find frequency components in an input IF signal.