Add files via upload
This matlab code plots range profile for 01 Tx antenna and 04 Rx antenna. The software extracts I and Q data from the radar cube used to store data in TI format. It performs non-coherent integration of chirps in time by performing 2D FFT and absolute summation across the antenna dimension. This code is a MATLAB implementation performed on actual data taken from ADC of AWR1642 with number of samples,1024. The implementation is based on the same method used in AWR1642 out of box demo in mmwave version 03_06_00_00-LTS